code refactoring and adding files to an additional task
