+++ controllers/auth/getCurrent.js
@@ -1,0 +1,9 @@
+const getCurrent = async (req, res) => {
+    const { email, password } = req.user;
+    res.json({
+        email,
+        password,
+    });
+};
+
+module.exports = getCurrent;
