--- controllers/auth/iindex.js
@@ -1,5 +1,0 @@
-const auth = require('./auth');
-
-module.exports = {
-    auth: ctrlWrapper(auth),
-};
