+++ controllers/auth/index.js
@@ -1,0 +1,14 @@
+const register = require('./register');
+const login = require('./login');
+const getCurrent = require('./getCurrent');
+const logout = require('./logout');
+const updateSubscription = require('./updateSubscription');
+const { ctrlWrapper } = require('../../helpers');
+
+module.exports = {
+    register: ctrlWrapper(register),
+    login: ctrlWrapper(login),
+    getCurrent: ctrlWrapper(getCurrent),
+    logout: ctrlWrapper(logout),
+    updateSubscription: ctrlWrapper(updateSubscription),
+};
