+++ controllers/contacts/listContacts.js
@@ -2,9 +2,15 @@ const { Contact } = require('../../models/contact');
 
 const listContacts = async (req, res) => {
     const { _id: owner } = req.user;
-    const { page = 1, limit = 20 } = req.query;
+    const { page = 1, limit = 20, favorite } = req.query;
     const skip = (page - 1) * limit;
-    const result = await Contact.find({ owner }, '-createdAt -updatedAt', { skip, limit }).populate(
+
+    const filter = { owner };
+    if (favorite) {
+        filter.favorite = favorite === 'true';
+    }
+
+    const result = await Contact.find(filter, '-createdAt -updatedAt', { skip, limit }).populate(
         'owner',
         'email'
     );
