+++ models/user.js
@@ -43,9 +43,14 @@ const loginSchema = Joi.object({
     password: Joi.string().min(6).required(),
 });
 
+const updateSubscriptionSchema = Joi.object({
+    subscription: Joi.string().valid('starter', 'pro', 'business').required(),
+});
+
 const schemas = {
     registerSchema,
     loginSchema,
+    updateSubscriptionSchema,
 };
 
 const User = model('user', userSchema);
