+++ routes/api/auth.js
@@ -1,6 +1,6 @@
 const express = require('express');
 
-const ctrl = require('../../controllers/auth/auth');
+const ctrl = require('../../controllers/auth');
 
 const { validateBody, authenticate } = require('../../middlewares');
 
@@ -15,5 +15,12 @@ router.post('/login', validateBody(schemas.loginSchema), ctrl.login);
 router.get('/current', authenticate, ctrl.getCurrent);
 
 router.post('/logout', authenticate, ctrl.logout);
+
+router.patch(
+    '/users',
+    authenticate,
+    validateBody(schemas.updateSubscriptionSchema),
+    ctrl.updateSubscription
+);
 
 module.exports = router;
